Created FM Index implementation using checkpointing and SA sampling
Access the FMIndex class by using "import proteinIndex"

The FMIndex class can build a new FM index given an input t, or it can read a version of itself saved to file (not tested yet). It can also query itself to find the offsets of patterns within the original string.